Rewrite of the sandbox with another approach, thread based.
Instead of "emulating" syscalls, the untrusted process now runs two threads. One
of them is running the untrusted code still under SECCOMP while the other
one is a trusted routine which will performs syscalls when the helper process
asks him.

This is in fact a generalization of the Chrome browser method.

There are still some TODO:

  - The security rules are still NOP
  - Performances penalty: Each untrusted syscall generates at least 6
    read/write (4 of them are dedicated to write the return values).
  - Memory area of the routine must never be writable for the untrusted
    thread. It was not investigate yet but maybe we should to not call the
    VDSO page anymore to really perform the syscalls because it pushes
    variables on the stack.