prototype done: trusted thread initialization is now PIC
Our goal is to no longer rely on LD_AUDIT feature. Initially, we wanted
to make the initialization inside the loader (ld-linux.so), just before
%eip is set to ELF entry point.

We first though "hey, why not write a minimalistic ld?" but we gave up because
we were afraid of GNU hacks which would be needed to handle GNU Libc.

So we thought "use the source Luke, and create a patch". But, we know this will
never been applied upstream and users will still have to recompile a loader which
is painful.

Instead, we prefer to patch ld's ELF binary. More to come...